feat: query jpql

+++ algafood-api/src/main/java/com/algaworks/algafood/domain/model/Restaurante.java
@@ -1,5 +1,7 @@
 package com.algaworks.algafood.domain.model;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -25,7 +27,7 @@ public class Restaurante {
 	private String nome;
 	
 	@Column(nullable = false)
-	private Double taxaFrete;
+	private BigDecimal taxaFrete;
 	
 	@ManyToOne
 	@JoinColumn(nullable = false)
